Rename various screen files + Update vector extension names
